candidate-backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
sqlalchemy>=2.0.36
python-dotenv>=1.0.1
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0
langchain-groq>=0.1.0
chromadb>=0.5.0
python-multipart>=0.0.12
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4